Moved projectfields functions into data utils
